@@ -1,11 +1,12 @@
 import 'package:flutter/material.dart';
 
+
 import 'form.dart';
 import 'form_field.dart';
 import 'form_row.dart';
 
 class TFormCell extends StatefulWidget {
-  TFormCell({Key key, this.row}) : super(key: key);
+  const TFormCell({Key? key, required this.row}) : super(key: key);
 
   final TFormRow row;
 
@@ -27,32 +28,28 @@ class _TFormCellState extends State<TFormCell> {
     } else {
       widget = TFormField(row: row);
     }
-    // backgroudcolor white default
     widget = Container(
-      child: widget,
       color: Colors.white,
+      child: widget,
     );
     // animation
     widget = row.animation ?? false
         ? TweenAnimationBuilder(
-            child: widget,
-            duration: Duration(milliseconds: 500),
-            builder: (BuildContext context, value, Widget child) {
-              return Opacity(
-                opacity: value,
-                child: child,
-              );
-            },
-            tween: Tween(begin: 0.0, end: 1.0),
-          )
+      duration: const Duration(milliseconds: 500),
+      builder: (context, value, child) {
+        return Opacity(opacity: value);
+      },
+      tween: Tween(begin: 0.0, end: 1.0),
+      child: widget,
+    )
         : widget;
     // divider
     widget =
-        TForm.of(context).divider != null && row != TForm.of(context).rows.last
-            ? Column(
-                children: [widget, TForm.of(context).divider],
-              )
-            : widget;
+    TForm.of(context).divider != null && row != TForm.of(context).rows.last
+        ? Column(
+      children: [widget, TForm.of(context).divider],
+    )
+        : widget;
     return widget;
   }
 }
