@@ -2,19 +2,19 @@ import 'form_row.dart';
 
 List formValidationErrors(List<TFormRow> rows) {
   List errors = [];
-  rows.forEach((TFormRow row) {
+  for (var row in rows) {
     if (row.validator != null) {
-      bool isSuccess = row.validator(row);
+      bool isSuccess = row.validator!(row);
       if (!isSuccess) {
         errors.add(row.requireMsg ?? "${row.title.replaceAll("*", "")} 不能为空");
       }
     } else {
       if (row.require ?? false) {
-        if (row.value == null || row.value.length == 0) {
+        if (row.value == null ) {
           errors.add(row.requireMsg ?? "${row.title.replaceAll("*", "")} 不能为空");
         }
       }
     }
-  });
+  }
   return errors;
 }
